+++ src/modules/Elsa.Workflows.Core/Activities/Flowchart/Activities/Flowchart.TokenCentric.cs
@@ -1,0 +1,9 @@
+namespace Elsa.Workflows.Activities.Flowchart.Activities;
+
+public partial class Flowchart
+{
+    private async ValueTask OnChildCompletedTokenBasedLogicAsync(ActivityCompletedContext context)
+    {
+        
+    }
+}
